feat(auth): implement JWT authentication and user-scoped job access

+++ server/controllers/jobControllers.js
@@ -2,7 +2,11 @@ const Job = require("../models/jobModel");
 
 const addJob = async (req, res) => {
   try {
-    const job = await Job.create(req.body);
+    const jobData = {
+      ...req.body,
+      userId: req.user.id
+    };
+    const job = await Job.create(jobData);
 
     res.status(201).send({ msg: "Job created." });
     console.log("Job Created", job);
@@ -14,7 +18,8 @@ const addJob = async (req, res) => {
 
 const viewJobs = async (req, res) => {
   try {
-    const jobs = await Job.find();
+    const userId = req.user.id;
+    const jobs = await Job.find({userId});
     res.status(200).send({ msg: "Jobs Fetched.", jobs: jobs });
     console.log("Jobs Fetched.");
   } catch (error) {
@@ -26,12 +31,13 @@ const viewJobs = async (req, res) => {
 const deleteJob = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.id;
 
     if (!id) {
       return res.status(400).send({ msg: "Job Id is required in URL." });
     }
 
-    const deleteJob = await Job.findByIdAndDelete(id);
+    const deleteJob = await Job.findByIdAndDelete({_id: id, userId});
 
     if (!deleteJob) {
       return res.status(404).send({ msg: "Job not found." });
@@ -48,9 +54,15 @@ const deleteJob = async (req, res) => {
 const updateJob = async (req, res) => {
   try {
     const { id } = req.params;
+    const userId = req.user.id;
 
     if (!id) {
       return res.status(400).send({ msg: "Job Id is required in URL." });
+    }
+
+    const job = await Job.findOne({_id: id, userId});
+    if (!job) {
+      return res.status(400).send({msg:"Unauthorized user."});
     }
 
     const updateJob = await Job.findByIdAndUpdate(id, req.body, {
